scraper/requirements.txt
request
BeautifulSoup4
PyMySQL